+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableBooleanArray.kt
@@ -218,6 +218,12 @@ public value class ImmutableBooleanArray @PublishedApi internal constructor(
             values.forEachIndexed(action)
 
     /**
+     * See [BooleanArray.all]
+     */
+    public inline fun all(predicate: (predicate: Boolean) -> Boolean): Boolean =
+            values.all(predicate)
+
+    /**
      * Leaves this immutable array as is and returns an ImmutableBooleanArray with all elements
      * sorted according to the natural sort order of the value returned by the [selector].
      *