+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableByteArray.kt
@@ -216,6 +216,11 @@ public value class ImmutableByteArray @PublishedApi internal constructor(
             values.forEachIndexed(action)
 
     /**
+     * See [ByteArray.all]
+     */
+    public inline fun all(predicate: (predicate: Byte) -> Boolean): Boolean = values.all(predicate)
+
+    /**
      * Leaves this immutable array as is and returns an ImmutableByteArray with all elements sorted
      * according to the natural sort order of the value returned by the [selector].
      *