+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableFloatArray.kt
@@ -217,6 +217,11 @@ public value class ImmutableFloatArray @PublishedApi internal constructor(
             values.forEachIndexed(action)
 
     /**
+     * See [FloatArray.all]
+     */
+    public inline fun all(predicate: (predicate: Float) -> Boolean): Boolean = values.all(predicate)
+
+    /**
      * Leaves this immutable array as is and returns an ImmutableFloatArray with all elements sorted
      * according to the natural sort order of the value returned by the [selector].
      *