+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableIntArray.kt
@@ -215,6 +215,11 @@ public value class ImmutableIntArray @PublishedApi internal constructor(
             values.forEachIndexed(action)
 
     /**
+     * See [IntArray.all]
+     */
+    public inline fun all(predicate: (predicate: Int) -> Boolean): Boolean = values.all(predicate)
+
+    /**
      * Leaves this immutable array as is and returns an ImmutableIntArray with all elements sorted
      * according to the natural sort order of the value returned by the [selector].
      *