+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableLongArray.kt
@@ -216,6 +216,11 @@ public value class ImmutableLongArray @PublishedApi internal constructor(
             values.forEachIndexed(action)
 
     /**
+     * See [LongArray.all]
+     */
+    public inline fun all(predicate: (predicate: Long) -> Boolean): Boolean = values.all(predicate)
+
+    /**
      * Leaves this immutable array as is and returns an ImmutableLongArray with all elements sorted
      * according to the natural sort order of the value returned by the [selector].
      *