+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableShortArray.kt
@@ -217,6 +217,11 @@ public value class ImmutableShortArray @PublishedApi internal constructor(
             values.forEachIndexed(action)
 
     /**
+     * See [ShortArray.all]
+     */
+    public inline fun all(predicate: (predicate: Short) -> Boolean): Boolean = values.all(predicate)
+
+    /**
      * Leaves this immutable array as is and returns an ImmutableShortArray with all elements sorted
      * according to the natural sort order of the value returned by the [selector].
      *